Add check to compare current and requested partition for local profiles
Adds a function to check whether the requested partition profile
matches the actual partition on which the nextflow head job is
running (i.e., the one on which the initial sbatch job was started).
The check is only performed for the local execution profiles and it
ensures that the resource thresholds defined for each profile are
applicable to the actual partition that is being used.

The check is performed by comparing the $SLURM_JOB_PARTITION environment
variable with the selected config profile.