Allow dynamic resource allocation for local execution profiles
Defines a function that grabs the allocated number of CPUs and RAM from
the current execution node and assign this value to the maximum
thresholds. This is done only for the *_local profiles, which are used
to run nextflow in local executor mode on a single slurm node.

Fallback values are defined corresponding to the total available
resources (corresponding to the request of an entire node).