Remove exitReadTimeout from individual profiles
Removes the executor.exitReadTimeout directive from the individual
profiles, to ensure that the executor scope in the main body of the
config is not ignored. See warnings here
https://nf-co.re/docs/usage/tutorials/step_by_step_institutional_profile#profiles-scope
and here
https://www.nextflow.io/docs/latest/config.html#config-profiles

The value of the timeout directive was set to that of the maximum
possible wall time (i.e., skylake partitions).