Fix unknown config attribute error
The host variable started throwing errors upon launch. The reason
was that the defined host variable was being called inside the
local scope of a function, where it was not accessible. To fix it,
either the partition_checker() function could be tweaked to take
an original parameter (String host), or the host variable could
just be omitted in favour of the underlying logic (= fetching the
content of an environment variable). The latter option was chosen.