Fix partition checker logic
The function that tests whether nextflow is running on the correct
slurm partition was broken; try-catch was used instead of conditionals
and a missing == comparison.
Now the code correctly warns the user if nextflow is running on a
login node, and aborts the run if the current partition does not
match the requested one (in the -profile flag).
It does not abort runs launched on the login node to allow easier
debugging.