Update checks for singularity/apptainer and nextflow environment variables
Improves checks for the environment variables for
singularity/apptainer/nextflow cache and tmp dirs:
- APPTAINER_TMPDIR/SINGULARITY_TMPDIR,
- APPTAINER_CACHEDIR/SINGULARITY_CACHEDIR or
- NXF_APPTAINER_CACHEDIR/NXF_SINGULARITY_CACHEDIR
Only the nextflow cache is set to a default value if it is not set in the
launch environment (via singularity.cacheDir). Before, this location
was always set and any value set in the env var was ignored.
The other two cannot be overriden in the env scope because that only affects
the task environment variables, not the environment in which nextflow is launched.
See: https://www.nextflow.io/docs/latest/config.html#scope-env
Warnings are issued for the user when any of these variables are not set.
Also clarifies code comments and updates documentation.