Simplify profiles through automatic partition selection
Simplifies the structure of this config by allowing the
appropriate values for each partition/queue to be selected
automatically.

The general profile `vsc_calcua` now defaults to the Slurm
scheduler. For local execution mode on a single node, a sub-profile
`single_node` needs to be appended.

Regardless of the type of executor, the current Slurm partition name
is read out from the environment variables. Based on this,
the correct settings (mem, cpu, time limits and queue) are applied.

For single node execution mode, the resource limits are still read
out dynamically. For Slurm execution, they are based on the node type.

The slurm profile that allowed tasks to be submitted across different
partiions was removed, because it was a hassle to debug and configure.
I.e. note for future attempts: are task.mem-type values re-read for
each task or only on launch of the profile?