Update munin.config
The line:
module    = { params.sentieon ? 'sentieon/202112.02' : null }
was changed to:
module    = { params.sentieon ? 'sentieon/202112.05' : null }

This is required by the Sentieon software having now come up with to a patched version due to their digital certificate being expired and thus renewed. We have set up the patched version  202112.05 now at BTB and we need Sarek to pick up this one instead of the old 202112.02, or it will no longer work with Sentieon.